requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
idna==3.10
paystackapi==2.1.3
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
